requirements.txt
# -----------------------------
# Core Framework
# -----------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0

# -----------------------------
# Google Gemini API (ONLY correct library)
# -----------------------------
google-generativeai==0.5.2

# -----------------------------
# Vector Database (Qdrant)
# -----------------------------
qdrant-client==1.9.1

# -----------------------------
# Data Models (Pydantic v2)
# -----------------------------
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-settings==2.2.1

# -----------------------------
# Utilities
# -----------------------------
requests==2.31.0
python-dotenv==1.0.1

# -----------------------------
# Optional (recommended)
# -----------------------------
numpy==1.26.4        # needed by some embed libraries

sqlalchemy==2.0.29
psycopg2-binary==2.9.9